Refactor gen-resume script

+++ scripts/gen-resume.sh
@@ -2,17 +2,24 @@
 set -euo pipefail
 
 PREFIX="Thanh-Ethan-Nguyen-resume"
+RESUME_DIR="resume"
 
 function main() {
   local version
+  version=$(jq -r '.resume_version' package.json)
 
-  cd resume/
-  version=$(cat version.tex)
+  if [[ -z "$version" ]]; then
+    echo "Missing version"
+    return 1
+  fi
 
+  cd "$RESUME_DIR/"
+
+  echo "$version" > version.tex
   local output="$PREFIX-$version"
+
   docker compose up resume
   cp "index.pdf" "../public/$output.pdf"
-
   docker compose down -t0
 }
